lib: improve flexibility of `args` param for `process`
Take `&[T]` where `T: AsRef<str>` instead of `&[String]` and
thus also support `&[&str]`.

This is not an API-breaking change, existing code will work with
it perfectly.

When building the first version of this library, the original idea
was to use `&[&str]` but the fact that Rust provides command line
args as `String` demanded a change to `&[String]`, not wanting to
force users to `collect()` args then build a respective array of
references. I was not then familiar with `AsRef`. Although most
users will likely work on args from the environment and thus
`&[String]`, this removes the unnecessary restriction from using
`&[&str]`.