Swap `Options` With `OptionSetEx` and `OptionSet` pair
`OptionSetEx` is the 'extendible' version, holding options in `Vec`s.
`OptionSet` is the 'non-extendible' version, which is still mutable,
to an extent, but holds references to slices of options.

`OptionSet` can be used for creation of static option sets.
`OptionSetEx` can be used when needing to dynamically modify the
option set.