lib: fix 'short-opt-same-arg-data' analysis bug
During the course of development of some significant upcoming
enhancements, an incorrect analysis was noticed with a random manual
test involving a short option taking a single char of
"in-same-argument" data. Investigation found further examples of
failings using multi-byte chars. Ultimately the cause was identified
as incorrect use of the `char_indices` method.

These example situations happened to not be covered in the test suite
and seem to have unfortunately been missed also previously in manual
testing prior to publishing the 1.0 release (but are covered by an
expanded test suite in pending work).

It seems that I mistook the integer returned by this method as being
equivalent to that supplied by the common iterator `enumerate`
method, i.e a counter, when in fact it is actually the byte index of
the char's location in the string.

The loop index naturally expected to run [0..n-1] when with ascii
input, due to skipping over of the single-dash prefix, this integer
ran [1..n], and thus was off by one. When given more than one char
of input data "in-same-argument", this worked fine, but with one char
only, it meant that it incorrectly determined the current short
option to be the last char in the argument and thus that this was an
"in-next-argument" situation, therefore consuming the next argument.
Furthermore, naturally no `break` occurred in the "in-next-argument"
branch, thus it then processed the final character in the argument as
a possible short option. This flaw naturally also causes things to go
wrong in multi-byte char situations.

Simple examples, for a data-taking short option 'o':

 1. `app -oa b`

    gave:
    { Short('o' -> "b"), [UnknownShort/Short/ShortMissingData]('a') }
    instead of:
    { Short('o' -> "a"), NonOption("b") }

 2. `app -oo b c`

    gave:
    { Short('o' -> "b"), Short('o' -> "c") ]
    instead of:
    { Short('o' -> "o"), NonOption("b"), NonOption("c") }

This fix implements correct use of the `char_indices` method, and was
able to simplify position tracking for "in-same-arg" data extraction
in doing so. It has been tested against the enhanced test suite of
the in-development next big release.