lib: fix broken ambiguity override
cherry-picked from later development branch

It was noticed that a `break` mistakenly inserted as an optimisation
during original development actually broken correct handling of
abbreviated matching. Specifically, if multiple abbreviated matches are
found, it would immediately break as ambiguous, instead of continuing to
look for an exact match which would override this, per specification. It
would only work correctly if the exact match was found prior to finding
two abbreviated matches, which depended upon the order of options in the
available options set.

The 1.0.x test suite does not cover this specific situation