lib/engine: fix rare abbreviated match inefficiency
If the option name was the same byte length as the candidate, but different,
then first the equality check would do a byte-by-byte comparison before
eventually rejecting, then the `starts_with` would start doing the same
before either rejecting or accepting. Thus any common prefix would get
processed twice.

This fixes this potential inefficient scenario by enforcing in the second
case that the length of the name cannot be equal to that of the candidate.
Either the length is equal, in which case only full equality is checked, or
it is less, in which case `starts_with` is checked.

An alternate solution would be to use a wrapping `starts_with` check, and
on success, check length to determine whether fully equal or abbreviated,
but `starts_with` uses `Pattern` based stuff which may be significantly
less efficient than the simple full equality check, so this solution was
used instead, even though an additional integer comparison is introduced.