add understanding of command arguments
While already compatible with programs using command arguments, the burden
was placed upon the user to seek out and recognise non-options as commands
and respond accordingly in handling options from a global set of options.

Adding understanding of commands significantly simplifies things for users
with minimal impact.

Users can now add a list of commands to a parser, each with their own
sub-option-set and sub-commands. Upon recongising a command in an argument
list, all subsequent arguments will be assessed against that command's
sub-option-set and sub-command set. This thus effectively partitions the
argument list per use of commands. Within each partition, the first
non-option is assumed to be a command argument, if a non-empty command
set is in play, and thus will result in a command if it matches one
(remaining a non-option otherwise).

There is no limit to sub-command nesting. This is a natural consequence of
how it was implemented; any restriction would be arbitrary.