lib/tests: fix expectations of flaw lists for dup checks
Only a single instance of each duplicate is added to the flaw
list, not `n-1` for `n` instances.

Note, the flaw sets are not currently used unless the `is_valid`
check incorrectly reports `true`, with their purpose then to
help compare the actual validation flaw set. This will be changed
in a later commit.