remodel on `Parser` struct
Preparation for the introduction of built-in command argument
handling.

The addition of this feature will see a new command set in
addition to the option set, and with each command optionally
having its own option set and sub-command set.

Initially adding the command descriptors into the existing option
set objects was tried, to avoid too much change, but logically
commands are not options, and it made a bit of a mess anyway.
Thus the introduction of this new `Parser` object, which takes
on the role of holding the `process` (now called `parse`) method,
and holds the option set and settings. In a follow up commit it
will be expanded to also hold the new command set descriptor data.