ci: bump minimum supported version to 1.31
The intention was to stick to 1.26 for a little while, supporting a few
releases back from the latest stable, but I ended up experiencing an ICE
on 1.30 during the latter part of working on the recent `OsStr` refactor
(merging of `str` and `OsStr` parsing), which then only went away with my
distro upgrading to 1.31. I can't be certain that the final commit (or
final version 2.0) will still fail on <1.31, but let's just bump to 1.31
and say to hell with it. This way we also get to jump on to those nice
new features now anyway :D