lib: support optional data values
I snubbed this for some reason with the original design, and later was
thinking that it would require peeking at the next argument to see if it
looked like an option and thus consider "optionally" consuming it, which
seemed flaky. Re-reading the getopt documentation, though, with an optional
data-taking option, the behaviour that was adopted for that was actually
just for data to only be consumed from the same argument.

This changes the `expects_data` option attribute to an enum offering
flag/data/optional-data choices, and with the latter, implements the same
behaviour as just described for getopt.