refactor all-in-one style analysis
The existing model of `Parser::parse()` returning an `Analysis` object
which in turn wrapped `ItemSet` objects, catering to both programs that
use a command-argument design and those that do not, was not ideal. There
was duplication, in documentation at least, between methods directly
available on `Analysis` (intended as a shortcut for non-command-based use),
and those on `ItemSet`. Also, users would be expected to know to only use
the data-mining methods directly available on `Analysis` if their program
did not make use of commands.

This new design adds `Parser::parse_cmd()` to join `Parser::parse()`, with
the latter returning an `ItemSet`, for use by non-command based programs,
and the former being for use with command-based programs.

It also removes the command name param from `ItemSet`, instead using a new
enum to capture this, with `Analysis` now holding a vector of instances of
that enum, rather than `ItemSet` directly. This simplifies use of `ItemSet`
in the non-command case, and it also is a neater representation (imo) of the
command-breakdown of the analysis - iterating over the vector held by the
analysis object now is an iteration over 'command' name instances and blocks
of other items that appear inbetween (captured within an `ItemSet`). It
perhaps better corresponds to the idea that a command name will be used
within one function, to determine the command-handling function to call,
and the next one or two (command/itemset/itemset+command) will be handled
within that command-handling function (command and its associated items
being dealt with in separate functions), thus better maybe to be split
into different enum instances, rather than part of the same object?

One final detail is that no longer is there always an entry representing any
items that come before the first command, when it would be empty.

Note that `Analysis` has been renamed `CommandAnalysis` to reflect its
more specific purpose now

no changes necessary for the bin code