lib/tests: split common parser into two
let's have one without commands, and one with

there are a bunch of tests with unknown-command items that we take the
opportunity to convert back to positionals by making use of the non-command
based parser. (they only have unknown-command items instead of positionals
because the introduction of the unknown-command type occurred against a
test parser with commands.