replace `&OptionSet` with `OptionSet` and similar for command sets
`OptionSet` is essentially a reference "class" of object; it is simply a
wrapper around a pair of slice references.

passing around and holding an `&OptionSet` just adds an extra pointer to
walk through when accessing the underlying lists, which is completely
unnecessary. Note that `OptionSet` should be cheaply copy-able.

fyi: this change came about from an issue implementing the trait
`OptionSetProvider` for generic option set sources (incomplete). some parts
of the code base demand a reference to a set, creating one to an empty
`OptionSet` where not provided. this was incompatible with the change to
generic option sets (impossible for trait method to give up a ref to a
static instance, where the instance is an "owned" type, e.g. holds a
`Vec`). it would have demanded an `Option` wrapper be added, which would
need unwrapping in the parsing engine. making this change though means that
we can avoid that (an implementation of the trait can, via a trait method,
give up an empty list object which is stored directly).