lib/engine: split out parser-data components
i.e. do not re-use the `Parser` type within the iterator objects

one benefit is that since the short option set iterator only makes use of
the option set, not the command set, nor any of the settings, it thus
shrinks in size, and since the main iterator holds a copy of it, that
shrinks too.

we could further optimise by only storing the short option set ref in the
short set iterator, rather than the entire `OptionSet`, but that's a
relatively small optimisation, and one that would need to be reversed by
the next big planned change.