lib/engine: do not cache arg assessor
caching which arg assessor to use (standard or alternate) offers a tiny
efficiency advantage in parsing, at the cost of a larger iterator object.

however, caching it has an impact on the design of setting-changing
methods. it is currently (since no longer holding settings in the short
option set iterator) the only thing holding back a redesign of the setting
read/write methods of the main iterator to simply give up a mutable ref to
the settings attribute, rather than having one method to get a copy of the
whole thing and another to overwrite the whole thing. (the issue is the
need to update the cached selected arg assessor).

this removes the caching in order to allow redesign of the setting changing
methods. (not that it would be common for anything to actually need to
change settings part way through parsing, it is thought, maybe with command
based parsing?)