lib: simplify changing iterator settings
now that we have 1) stopped using `Parser` within the iterator objects and
thus the short option set iterator no longer holds settings, and 2) we have
stopped caching a function pointer for the argument type assessor, we were
now free to simplify the methods for changing settings to one that just
gives up a mutable ref to the settings attribute.

this means that individual settings can now just be directly changed,
rather than having to copy the entire structure back and forth.