split parser into two
so that we now have `Parser` and `CmdParser`

benefits:
 - no need for blocking mistaken use of `parse` instead of `parse_cmd`
   for command-based programs with assertion.
 - simplifies choice of available methods to two instead of three.
 - helps separate command related stuff from non-command based uses

another benefit is that it makes even more sense with a possible further
change that moves command handling to a new layer on top of the main
iterator, which itself is prompted by issues relating to mess of lifetimes
surrounding implementation of an even greater enhancement. To reduce the
mess of lifetimes we end up wanting to split out the command based stuff
to a new layer, which leaves us with two different iterator objects, which
then leaves us with four different parsing methods, two for commands and
two not, so it ends up making sense to split up the parser then anyway.

unfortunately:
 - `Parser` still needs the `commands` attribute, since the engine does not
   know the difference between the parsers. this can only be solved by
   splitting command handling out to a layer on top of the main iterator.
 - the `parse_iter` and `parse` methods of `CmdParser` are having to
   convert self to a [`Parser`] in order to create a ref for creation of a
   `ParseIter`.