introduce `Args` based argument capture
the reasons for wanting this is covered thoroughly in the associated
documentation

note, regarding the test suite, the `arg_list` macro was changed to no
longer create `OsStr` types with the given strings, since these are not
considered to have static lifetimes unlike string literals, and thus it
would require two steps to create an arg list if this remained, we would
first need to create the array, and then secondly create the `Arg` object,
and no this cannot be done in one macro call while using the `Arg` is done
outside of it. This really isn't a big deal though since the engine still
sees them as `OsStr` anyway via `AsRef<OsStr>`, and there are some tests
still specifcally using `OsStr`.