stop serving arg index with items
cache it in the iterator and make available through a method, thus people
can still retrieve it if they want it (unlikely), but leaves things cleaner
for the majority who likely are uninterested.

note that in adding it to the main iterator we actually get to remove it
from the short-option-set iterator, thus no difference in iterator size.

wrt. `ItemSet` it was decided to bundle the index with items in a tuple
for now at least, to retain the same information as before.