stop serving data-location with items
cache it in the iterator and make available through a method, thus people
can still retrieve it if they want it (unlikely), but leaves things cleaner
for the majority who likely are uninterested.

note that:
 - we do **not** reset it to `None` on every short opt set iteration, since
   handling a data-value only occurs with the very last option in the set,
   so no point.
 - we currently have to cache it in both the main iterator and the short
   option set one, copying from the latter to the former on each iteration
   of the latter, which is not ideal. this can be improved later.
 - we are not currently integrating this with the 'indexed' iterators, that
   can occur separately.

again, like with the removal of arg indexes, this information is now
bundled into the data-mining oriented collection objects in a tuple form.