bundle data-location in indexed iterator tuple
when this attribute was moved out of items and instead cached in the
iterator for optional retrieval, like done with argument indexes, we did
not then bundle it into the tuple returned by the 'indexed' wrapping
iterator, leaving it to optionally do separately, which we are now.

we also temporarily had to add retrieval methods to those 'indexed'
iterator types such that the users of the 'indexed' iterators could get
access to this attribute in the meantime. here we now remove those
temporary methods.