lib/engine: simplify
it has never been ideal that the `ShortSetITer` type had to clone the arg
iter and opt set and recently also `last_data_loc`, but the `Iterator`
implementation constrained the interface of its private `do_next` method.
furthermore it could not hold a reference to its parent, as this would
introduce a new lifetime on the parent itself, and nor could it use a raw
pointer since we would then have cloning issues.

now that we have gotten rid of the unnecessary `Iterator` impl we have
made it possible to simplify things. now we just pass along a ref of the
parent when that makes each call to the short-set-iter `do_next()` method
and that now can simply use the parent attributes as needed (use the option
set; consume the next arg wrt. data values; and cache the last data value
location).