lib: ditch indexed data from data-mining objects
it is ignored in most data-mining method output, which these objects are
mostly useful for. It is expected that it is rare that anyone will actually
want these attributes anyway.

this does impact the test suite though, with that no longer being able to
check these attributes are correct, since it is based upon those objects.
we need to consider this further, whether to be satisfied with ignoring
those attributes for most tests, and maybe adding some iterative based ones
specifically to check correctness of them, or perhaps switch most tests
over to being iterative based, with some tests of generating correct
data-mining oriented objects.