add capability of reporting unexpected-positional
previously, since the number of accepted positionals in programs can vary
depending for instance on use of options, it was felt that doing anything
other than just accepting an unlimited number of positionals fell into the
same area as option relationships which would require a complex framework
to support and which would be most efficiently and simply expressed just in
application specific logic.

however I have reconsidered this position with respect to positionals,
since it is not ideal that users of data-mining objects for instance could
very easily overlook the presence of positionals by failing to check the
positionals related method.

here we implement a mechanism, based on simply counting the number of
positionals encountered and comparing with a given maximum. if above the
maximum then the new unexpected-positional problem item is returned, thus
this triggers the `problems` indicator of data-mining objects to flip to
`true` and this problem will be returned by the problem reporting
method(s). thus now users will no longer simply fail to notice them as
described above, so long as they set a correct expected quantity.

we retain the ability of course to specify the quantity as unlimited. the
`cargo run` command would need this for instance, which passes all
positionals on to the program it runs. note that this quantity can be
specified per-command.

note that this is a 'maximum' quantity, which helps with flexibility
somewhat.

in terms of dynamically adjusting the max dynamically in response to use
of options, a method is available on the iterator to allow updating it.
how to achieve the same with data-mining objects though is unresolved.

note, the design does not truly respect "unlimited" in terms of the
unlimited policy; the positionals counter increment will cause a
crash once u16::MAX positionals are reached, but this is acceptable
since that's a crazy number of positionals. Perhaps we should have the
parser terminate parsing at that number of arguments to prevent a
crash, maybe returning a problem type dedicated to this? Worth it to
redesign to remove this limit (the natural limit with Rust code would
be usize::MAX, not sure about OS limits)?