lib/tests: partial conversion to iterator based testing
after the iterator based parsing refactor the test suite remained using the
data-mining oriented analysis objects for the basis of the tests, with a
few extra tests specifically targeting the iterative model. this was
partially because as it was it encompassed both the iterative handling
(used in the background for constructing the data-mining objects) and the
data-mining object construction. it was also partly because we had separate
error and warn attributes back then and it was helpful to check that they
were correct in each test situation. another reason was because a solution
had yet to be implemented for adjusting the object to work with command
partitioning, and then the first such solution built this into the single
object available at the time, shared between command and non-command based
scenarios, and so we needed to be sure that this partitioning worked
correctly in all scenarios; a later revision split it into two separate
objects.

now though we have just redesigned things to move argument index and
data-location data out of the normal item variants, making it extra data
provided by an 'indexed' iterator on top of the normal iterator, and then
we have gone a step further in removing this data from the data-mining
objects, as likely unwanted by anyone there. this has meant that this data
now goes untested in the test suite, when really we want it tested in each
of the majority of tests. in fact an earlier revision of these commits
actually contained a bug which was hidden by the fact that the test suite
was not covering this data, proving the benefits of checking it.

furthermore, at this point, there is considerably little value in checking
the correctness of the data-mining object construction in each and every
test.

thus, here we move the non-command oriented set of tests to be iterator
based, which actually simplifies things somewhat (no need for the `Actual`
and `Expected` wrappers or `ItemSet` problem printer function.

however, we keep the `ItemSet` based stuff within the `analysis` test
module, since that is entirely focused upon data-mining. the tests there
now check both iterative results and then the construction of the
data-mining object, alongside then testing actual data-mining.

note that the `expected` macro has changed to now simply compile an array
for iterative tests, and the old macro has been renamed `dm_expected`.