lib: differentiate between no and empty string data where optional
with optional data-taking options (which can only take data in-same-arg),
we should be distinguishing between no data and empty string data, i.e.
the difference between `--foo` and `--foo=`.

in the former case we take this as being no value having been provided,
while in the latter we take this as being an empty string value being
provided.

it is thus important that we add this distinction to allow users to supply
empty string values, just as with mandatory data-taking options.

of course we can only make such a distinction with long options, it is
simply just not possible to do so with short options.

note that we have to avoid setting the cached data-location value in this
case; since the returned item is indistinguishable from a flag type item,
and since this is what is now re-used in describing expected items in
tests, we need to have data-location be `None` for tests to actually pass
(once testing of that attribute is restored). this is in a way more
consistent with the missing-data variant.

note also that this has an impact upon data-mining methods:
 - `get_last_value()`: previously, with an empty string value being
   returned, this would be what was reported by this function, if it was
   the last instance; now though the item will be ignored since the data
   value is `None` rather than `Some(_)`. We need to consider how exactly
   we should treat missing data values with respect to this function.
   Should it be expected that we ignore it and go back to the last actual
   value supplied; or should it be expected that we respect it to thus
   indicate that use without a value was the last use of such an option
   (maybe signalling that the default value should be used).
 - `get_all_values()`: similarly, the empty value was previously included
   here, but now is ignored. what to do? should we add an option wrapper?
would adding new options, or a boolean param be best?

furthermore note that `None` vs `Some(_)` for the data value attribute
of long and short item variants are now no longer reliable indicators of
whether the option is of flag type or data-taking type.