lib/engine: fix consistency with no optional-data
long options now correctly distinguish between no-data and empty string
values for optional data-taking options. we could not do the same for
short options, since there is no way to tell the difference.

here we switch from returning an empty string to `None` for short options,
for consistency - we will consider the data missing.

consider that with long options `--foo` is missing data, while `--foo=` is
one where we consider the user to be deliberately supplying an empty string
(just as the latter means the same with mandatory data-taking options); so
let's consider `-f` as missing, not forced empty string.

note that we drop setting the data-location value, just as with the related
change to long options in the previous commit.