bin: fix optional no-data analysis
with the change of optional data-taking options now having a data attribute
of `None` in their item variant rather than an empty string, we lost the
ability to distinguish between flag type and data-taking type items from
simply looking at the `Option` variant of the item's data attribute.

it seems nicer however in this test binary to distinguish the difference
in the output, which this change enables.