lib/engine: restore data-loc for no-optional-data items
this gives users a hint for telling the difference between a flag type
option match and a optional-data-taking option match with no data, with
both being of the form `Item::Long(_, None)` or `Item::Short(_, None)`,
helping avoiding having to look up the option to check its type.

i.e. if the data-location attribute is `None` then this means that the
match came from a flag type option match, whereas if it is
`Some(DataLocation::SameArg)` then it is from the other.