lib: change positionals default from unlimited to zero
i.e. by default all positionals become unexpected-positional problem item
variants. they will only be accepted as non-problem positionals if the
user explicitly changes the requirements to allow one or more in their
parser.

this is a much better default to take, now that we can offer it, such that
we are not unexpectedly accepting stuff. there would not be a concern with
iterative parsing, since the user would have to explicitly react to a
positional, but with data-mining, we would not want to just silently
collect positionals that are not wanted just because the user forgot to
change the requirements.