lib/tests: add and use new `dm_item` macro
this macro is now dedicated to creating data-mining object oriented items

the need for it is as an alias, with potential for data-mining objects to
drop the extra 'index' data they currently still hold after its extraction
from within item variants. it will allow switching the tests over to be
easier, since we just change the `dm_item` macro to ignore the index
attributes.