rename optional-data-taking type
`OptionalData` and `@opt_data` were too clunky compared to `Data`/`@data`
and `Flag`/`@flag`. Even talking about an "optional data-taking option" is
so also. This option type can be thought of as a flexible one which can be
used either in flag form or (with restriction) data-providing form.

Thus it is cleaner to use the term 'flexible', or as chosen here 'mixed'
instead.