lib: add the `OptionPair` type
This is *was* intended to be an enum like `FindOption` except additionally
carrying the option type, however the capabilities of `const` functions is
so limited that my plan for that fell flat, so this was the alternative.

This new type is one for expressing a related pair of option identifiers
along with their option type, together in one structure. It comes with
`const` function methods for generating both long and short specific
option types as well as `FindOption`.

The purpose of having this is so that you can define the pair with a single
declaration of their type, helping avoid mistakes, and you can, through
the methods just mentioned, spawn objects for option set creation and data
mining use, again reducing mistakes. Importantly, with these methods being
`const`, this does not impact the ability to create static option sets.

Thus for instance:
```
const HELP_OPT: OptionPair = optpair!(@flag 'h', "help");
static OPTIONS: OptionSet = option_set!(
    @long [
        HELP_OPT.as_long(),
    ],
    @short [
        HELP_OPT.as_short(),
    ]
);
//...
if analysis.option_used(HELP_OPT.to_findopt()) {
    println!("{}", HELP_TEXT);
    return;
}
```