lib: simplify option types
having labels for each attribute is overly verbose for these

for a few places where we want clarity and in one case where we have an
assert where type checks would not point out a problem referring to the
wrong attribute if something later changed, we use simple inline
functions to get the attributes. i.e. the long and short types (not the
pair type) have `ident()` and `ty()` methods