lib: extract command name issues in data-mining object
the `CommandAnalysis` data mining object already breaks up items by use of
commands, extracting command items up to the level of item groups. this
extends this to also doing the same with unknown and ambiguous command
items.

the purpose of breaking items up into groups like this is such that it
corresponds to how commands partition the argument list, and with every
partition there is a command and/or a related item set to process, and
data-mining can be performed on just the group of items relating to each
command.

it makes sense to start a new group of items after failed command name
matches as well as good matches. all items before the failed match are
associated with the previous command (if any), while all those that occur
after it are associated with the failed match (or have been
misinterpreted). of course in a stop-on-problem situation no further items
would have been collected.

the only issue is should such an item have been incorrectly identified as
a failed command name match instead of a positional (a topic which needs
more thought), which would then result in three components in the analysis
which would need using as one (set; failed-match; set). We could perhaps
address this of course with a method that changes the failed match into a
positional and combines the three, re-evaluating the `problem` attributes.
re-evaluating with respect to the positionals policy would also be needed.
the core aspect to this though is relevant irregardless of having split
things up.