lib: add new setting to control return of early terminator item
most programs will not actually care for an actual item being returned for
an early terminator, so this new setting allows it to be filtered out.

previously with iterative parsing users would be forced to have a match
block for the early terminator which did nothing, or let it be caught by a
catch all that did nothing (`_ => {}`). if they mistakenly directed it to
`unreachable!()` or something because they were not using it then their
program would crash if a user used an early terminator. this setting now
means that it is responded to cleanly without bothering the user with an
actual item in the returned set, and thus users can freely direct it to
`unreachable!()` if they do not wish to actually do something special in
response to it.