lib: address issue with data mining mixed option values
as expressed in the removed todo notes: we have merged flag and data-taking
item variants together and furthermore implemented a means of
distinguishing `--foo` from `--foo=` for mixed type options, where the
former is a non-value and the latter an empty string value, and thus the
non-value instances are indistinguishable from flag type matches except
when looking at the data-location attribute (not recorded in data-mining
objects anymore). this has meant that non-value uses were getting ignored
by the `get_last_value` and `get_all_values` methods, which is not
necessarily what users may want.

there is no obviously correct answer as to whether users would want such
values to be ignored like this or otherwise. the only good solution was to
offer users choice with the addition of alternate methods.

alternatives considered:
 - taking a firm stance to ignore such instances with those methods
 - adding the extra `Option` wrapper to the result type as done with the
   new alternate methods added here and treating such instances the same as
   if data were given
 - adding the extra `Option` wrapper as just described but including a new
   boolean parameter to control whether or not to ignore such instances

trying to incorporate changes to the existing methods would have made usage
for non-mixed options and where the user does want to ignore such instances
considerably messier and less efficient.