lib: resolve issue of unknown-command vs. positional
a lingering issue in the design of this library at this point was what to
do in the circumstance of having a non-option argument that did not match
an available command, when commands are available; should this be reported
as an unknown command or a positional?

originally when commands were introduced, we just went with positional,
but later this decision was changed and the unknown-command item variant
introduced. however although this improved things in some respects, it was
imperfect, for instance in requiring users to fetch the original OsStr from
the original argument set if wanting to re-interpret as a positional
(back when unknown-command held a `Cow<str>`).

the recent introduction of positionals policy impacts this topic, but
initially no opportunity was available to properly consider how to best
handle the problem. it is even more important to pick the right result now
since it impacts the correct serving of positional item variants.

we would not want to introduce an arbitrarily chosen restriction which
makes this library incompatible with some application CLI designs. how I
saw the problem:
 - if no commands are available, then clearly the non-option is a positional
 - if command are available and it matches one, then it is a command
 - if no match and no positionals allowed per policy (only recently
   possible to express) then definitely an unknown-command
 - but if no match with the available commands and positionals are to be
   accepted, then what??
it was difficult to reason whether or not the final scenario was even
realistic, and lack of opportunity to properly clear my mind and think this
through stalled things in this area.

would a new per-command setting be required to give users choice over
which of unknown-command or positional the non-option would come out as?

having now been able to think it through, I realise that the answer is
actually quite simple (now having the positionals policy). the positionals
policy itself is giving the necessary information: we have a set of
commands available; if a non-option does not match one then:
 - if positionals policy MAX == 0, then go with unknown-command
 - otherwise go with positional
typically an application may use MAX==0 when there is a set of commands to
choose from, and possibly !=0 when not. if at a point where commands are
available the MAX is !=0 then this is the application designer explicitly
choosing to accept positionals as an alternative to selecting a command
and thus in this case only we should be taking it as a positional instead
of an unknown command.

thus (per argument):
 commands? no  -> positionals? no  -> unexpected-positional
                               yes -> positional
           yes -> match? yes -> command
                         no  -> positionals? no  -> unknown-command
                                             yes -> positional