combine long and short match items into one
instead of having separate item variants representing long and short option
matches, combine them into one, using the `OptID` enum for representing the
different identifiers.

we retain the ability to specifically pattern match a long OR short in a
single match arm, if we want to, but possibly at a slightly higher cost due
to two attributes now needing to be compared (depending upon how that
pattern match is implemented in assembly), however on the other hand we can
now have fewer match arms. it also enables cleaner code, and handling of
both long and short items within the same arm when comparing the identifier
within the arm's block.