add missing-positionals problem reporting
building upon the unexpected-positional support, here we extend it to allow
expressing 'fixed', 'minimum' and 'min-max' requirements, which translate
to a new missing-positionals problem item variant being issued if too few
positional arguments are supplied.

note, this changes the handling of the unlimited policy to actually no
longer be truly unlimited; it is now equivalent to `Max(u16::MAX)`, since
the policy now gets reduced down to simply (min, max) where u16::MAX is
used for max with `Unlimited` and `Min(_)` policies.

since u16::MAX (65535) is a crazy number of input arguments (imo), this is
not expected to be a problem.