combine short/long missing data problem item types
same benefits as just done with known option matches

this allows us to handle the base problem with a single match block arm.
the only issue is correctly printing the identifier with the correct
prefix, if wanting to include it in the output, but since that relies upon
knowing the option mode setting we cannot make any assumption by handling
it in a `Display`/`Debug` implementation.