combine short/long unknown option problem item types
same benefits as just done with known option matches

as demonstrated in the examples, this enables us to handle unknown options
of both long and short forms within the same match arm. we do not
necessarily need to print the relevant option prefix in the error message,
the identifier alone can do, and so this allows us to handle this problem
more cleanly.

note, due to unknown-long being an &OsStr we could not use the existing
`OptID` type as it was, we needed to either introduce a new wrapper or, as
we have done here, make it generic. this required an explicit implementation
of `Copy` and `Clone` since a `?Sized` implementation cannot be derived it
seems.

introducing a new `OptID` type would have been cleaner, but do we really
want a new type in the interface just for this one use case?